Export dataframe labels to excel

export_dataframe_labels exports the variable labels of a dataframe, which are created by labelled::var_label, to an excel sheet.

Usage

export_dataframe_labels(
  data,
  workbook,
  sheet_name = format(Sys.time(), "%d%b%Y_%H%M")
)

Arguments

data

(data.frame) Your dataframe whose variable labels you want to export to excel

workbook

(character) File path of excel workbook

sheet_name

(character) Name of excel sheet to export labels to

Defaults to system time at execution if none provided

Value

None

Examples

library(labelled)

df <- data.frame(
  column_a = sample(letters, 3),
  column_b = runif(3, 1, 10)
)
labels <- list(
  column_a = "This is column A",
  column_b = "This is column B"
)
var_label(df) <- labels

# Export column labels to an excel file in temp directory
output_filepath_1 <- file.path(tempdir(), "file1.xlsx")
export_dataframe_labels(df, output_filepath_1, sheet_name = "my_sheet")
print(list.files(tempdir(), pattern = "\\.(xlsx|xls)$"))
#> [1] "file1.xlsx"

# Delete the excel file in temp directory
unlink(output_filepath_1)
print(list.files(tempdir(), pattern = "\\.(xlsx|xls)$"))
#> character(0)