Extend a paired table

extend_paired is a wrapper around arsenal::paired that adds a marker which enables the joining of two paired tables together.

Usage

extend_paired(paired_obj, labels = NULL, bold_var = TRUE)

Arguments

paired_obj

(paired object) An arsenal::paired object

labels

(list) An optional named list of variable labels, for the data passed into paired_obj. labels may be set with labelled::var_label

bold_var

(logical) If TRUE, variable sub-headers will be wrapped with **

Value

(data.frame) A paired table appended with marker

Examples

library(arsenal)
library(tidyverse)
#> ── Attaching core tidyverse packages ──────────────────────── tidyverse 2.0.0 ──
#> ✔ dplyr     1.1.4     ✔ readr     2.1.5
#> ✔ forcats   1.0.0     ✔ stringr   1.5.1
#> ✔ ggplot2   3.5.1     ✔ tibble    3.2.1
#> ✔ lubridate 1.9.3     ✔ tidyr     1.3.1
#> ✔ purrr     1.0.2     
#> ── Conflicts ────────────────────────────────────────── tidyverse_conflicts() ──
#> ✖ dplyr::filter()      masks stats::filter()
#> ✖ lubridate::is.Date() masks arsenal::is.Date()
#> ✖ dplyr::lag()         masks stats::lag()
#> ℹ Use the conflicted package (<http://conflicted.r-lib.org/>) to force all conflicts to become errors
library(labelled)

# Generate fake patient data for illustration -------------------------------
generate_fake_patient_data <- function(id) {
  data.frame(list(
    patient_id = id,
    time = c("baseline", "12-month"),
    hba1c = runif(n = 2, min = 4, max = 14),
    ldl = sample(c("good control", "bad control"))
  ))
}

set.seed(2024)
data <- map(
  sprintf("P%s", str_pad(1:30, width = 2, pad = 0)),
  generate_fake_patient_data
) %>%
  bind_rows()
print(head(data, 10))
#>    patient_id     time     hba1c          ldl
#> 1         P01 baseline 12.369425 good control
#> 2         P01 12-month  7.208675  bad control
#> 3         P02 baseline  8.570092 good control
#> 4         P02 12-month 11.014203  bad control
#> 5         P03 baseline 12.765881  bad control
#> 6         P03 12-month  5.190482 good control
#> 7         P04 baseline 11.035531  bad control
#> 8         P04 12-month  9.157855 good control
#> 9         P05 baseline  5.103111  bad control
#> 10        P05 12-month 12.632152 good control

# Create paired table without labels ----------------------------------------
paired(time ~ hba1c + ldl,
  data = data %>% mutate(across(
    .cols = time,
    .fns = ~ fct_relevel(.x, "baseline", "12-month")
  )),
  id = patient_id,
  control = paired.control(
    digits = 2,
    numeric.stats = c("Nmiss", "meansd", "medianq1q3", "range"),
    numeric.test = "signed.rank",
    signed.rank.correct = FALSE
  )
) %>%
  extend_paired()
#>             Variable    baseline (N=30)    12-month (N=30)  Difference (N=30)
#> 1          **hba1c**                                                         
#> 2       -  Mean (SD)        8.45 (3.09)        8.76 (3.09)        0.31 (4.78)
#> 3 -  Median (Q1, Q3) 7.81 (5.80, 11.30) 8.66 (6.05, 11.78) 0.29 (-2.44, 3.74)
#> 4           -  Range       4.01 - 12.77       4.21 - 13.73       -7.58 - 9.12
#> 5            **ldl**                                                         
#> 6     -  bad control         14 (46.7%)         16 (53.3%)        14 (100.0%)
#> 7    -  good control         16 (53.3%)         14 (46.7%)        16 (100.0%)
#>   p value                   marker
#> 1   0.730                    hba1c
#> 2               hba1c -  Mean (SD)
#> 3         hba1c -  Median (Q1, Q3)
#> 4                   hba1c -  Range
#> 5   0.855                      ldl
#> 6               ldl -  bad control
#> 7              ldl -  good control

# Create paired table with labels -------------------------------------------
## Add labels
var_label(data) <- list(
  hba1c = "Hemoglobin A1c",
  ldl = "LDL control"
)

## Create paired table
paired(time ~ hba1c + ldl,
  data = data %>% mutate(across(
    .cols = time,
    .fns = ~ fct_relevel(.x, "baseline", "12-month")
  )),
  id = patient_id,
  control = paired.control(
    digits = 2,
    numeric.stats = c("Nmiss", "meansd", "medianq1q3", "range"),
    numeric.test = "signed.rank",
    signed.rank.correct = FALSE
  )
) %>%
  extend_paired(labels = var_label(data))
#>             Variable    baseline (N=30)    12-month (N=30)  Difference (N=30)
#> 1 **Hemoglobin A1c**                                                         
#> 2       -  Mean (SD)        8.45 (3.09)        8.76 (3.09)        0.31 (4.78)
#> 3 -  Median (Q1, Q3) 7.81 (5.80, 11.30) 8.66 (6.05, 11.78) 0.29 (-2.44, 3.74)
#> 4           -  Range       4.01 - 12.77       4.21 - 13.73       -7.58 - 9.12
#> 5    **LDL control**                                                         
#> 6     -  bad control         14 (46.7%)         16 (53.3%)        14 (100.0%)
#> 7    -  good control         16 (53.3%)         14 (46.7%)        16 (100.0%)
#>   p value                   marker
#> 1   0.730                    hba1c
#> 2               hba1c -  Mean (SD)
#> 3         hba1c -  Median (Q1, Q3)
#> 4                   hba1c -  Range
#> 5   0.855                      ldl
#> 6               ldl -  bad control
#> 7              ldl -  good control