Extend a tableby table

extend_tableby is a wrapper around arsenal::tableby that adds the following to a tableby table:

Marker for enabling the joining of two tableby tables together

Standardised mean difference (SMD) between the groups

Usage

extend_tableby(tableby_obj, data, smd = TRUE, smd_digits = 2, bold_var = TRUE)

Arguments

tableby_obj

(tableby object) An arsenal::tableby object

data

(data.frame) Data passed into tableby_obj

smd

(logical) If TRUE, computes the standardised mean difference (SMD)

smd_digits

(numeric) Number of decimal places to express SMD

bold_var

(logical) If TRUE, variable sub-headers will be wrapped with **

Value

(data.frame) A tableby table appended with marker and SMD

Examples

library(arsenal)
library(labelled)
library(MatchIt)
data(lalonde)

# Add labels (optional) -----------------------------------------------------
var_label(lalonde) <- list(
  age = "Age",
  educ = "Education",
  race = "Race",
  married = "Marital status"
)

# Create tableby table ------------------------------------------------------
tableby(treat ~ age + educ + race + married,
  data = lalonde,
  control = tableby.control(numeric.stats = c("Nmiss", "meansd", "medianq1q3", "range"))
) %>%
  extend_tableby(lalonde)
#>              Variable               0 (N=429)               1 (N=185)
#> 1             **Age**                                                
#> 2        -  Mean (SD)         28.030 (10.787)          25.816 (7.155)
#> 3  -  Median (Q1, Q3) 25.000 (19.000, 35.000) 25.000 (20.000, 29.000)
#> 4            -  Range         16.000 - 55.000         17.000 - 48.000
#> 5       **Education**                                                
#> 6        -  Mean (SD)          10.235 (2.855)          10.346 (2.011)
#> 7  -  Median (Q1, Q3)  11.000 (9.000, 12.000)  11.000 (9.000, 12.000)
#> 8            -  Range          0.000 - 18.000          4.000 - 16.000
#> 9            **Race**                                                
#> 10           -  black              87 (20.3%)             156 (84.3%)
#> 11          -  hispan              61 (14.2%)               11 (5.9%)
#> 12           -  white             281 (65.5%)               18 (9.7%)
#> 13 **Marital status**                                                
#> 14       -  Mean (SD)           0.513 (0.500)           0.189 (0.393)
#> 15 -  Median (Q1, Q3)    1.000 (0.000, 1.000)    0.000 (0.000, 0.000)
#> 16           -  Range           0.000 - 1.000           0.000 - 1.000
#>              Total (N=614) p value                     marker  SMD
#> 1                            0.011                        age 0.24
#> 2           27.363 (9.881)                   age -  Mean (SD)     
#> 3  25.000 (20.000, 32.000)             age -  Median (Q1, Q3)     
#> 4          16.000 - 55.000                       age -  Range     
#> 5                            0.633                       educ 0.04
#> 6           10.269 (2.628)                  educ -  Mean (SD)     
#> 7   11.000 (9.000, 12.000)            educ -  Median (Q1, Q3)     
#> 8           0.000 - 18.000                      educ -  Range     
#> 9                          < 0.001                       race 1.70
#> 10             243 (39.6%)                      race -  black     
#> 11              72 (11.7%)                     race -  hispan     
#> 12             299 (48.7%)                      race -  white     
#> 13                         < 0.001                    married 0.72
#> 14           0.415 (0.493)               married -  Mean (SD)     
#> 15    0.000 (0.000, 1.000)         married -  Median (Q1, Q3)     
#> 16           0.000 - 1.000                   married -  Range