Extract latest reading in specified window period of observation

Suppose that you have the following, extract_reading_from_interval will extract the latest laboratory reading for each patient within the respective window periods to be considered as the "baseline" reading, "3-month" reading etc.

a set of longitudinal data e.g. laboratory readings of patients

a set of parameters that define the window period of observation for considering readings that would constitute as "baseline" reading, "3-month" reading etc. for example:

baseline: T-180 days to T

3-month: T+31 days to T+90 days

where T is the index date which could be the enrolment date or date of first visit

Usage

extract_reading_from_interval(
  df_master,
  df_master_fields,
  df_long,
  df_long_fields,
  params
)

Arguments

df_master

(data.frame) Master dataframe to join the readings unto. df_master must minimally contain the following 2 columns:

A column of IDs to identify each patient

A column of index dates (what we refer to as T in the description above) for each patient

Both these columns will be identified in df_master_fields

df_master_fields

(list) A named list to identify the ID and index date columns in df_master

It takes on this format: list(ID = <your_ID_column>, index_date = <your_index_date_column>)

For example: list(ID = "patient_id", index_date = "enrolment_date")

df_long

(data.frame) Longitudinal data e.g. laboratory readings

df_long_fields

(list) A named list to identify the reading date and reading(s) in df_long

It takes on this format: list(reading_date = <your_reading_date_column>, reading = <your_reading_column(s)>)

reading_date (character) The name of the dataframe column that contains the dates of the readings

reading (character or character vector) The name(s) of the dataframe column(s) that contains the readings

For example: list(reading_date = "calendar_date", reading = c("bp_systolic", "bp_diastolic"))

params

(list) A named list where the names are the timepoints and the values are numeric vectors that represent c(lower_bound, upper_bound)

For example: params <- list(baseline = c(-180, 0), "3mth" = c(31, 90), "6mth" = c(120, 180))

Both lower and upper bounds are included in defining the window period i.e. lower_bound <= reading_date <= upper_bound

For params given above, extract_reading_from_interval will consider

Baseline reading: latest reading in T-180 days <= reading_date <= T window

3-month reading: latest reading in T+31 days <= reading_date <= T+90 days window

6-month reading: latest reading in T+120 days <= reading_date <= T+180 days window

Value

(data.frame) df_master with longitudinal readings joined unto

Examples

library(dplyr)

# Generate fake patient data for illustration -------------------------------
generate_fake_patient_data <- function(id) {
  data.frame(
    list(
      patient_id = id,
      enrolment_date = Sys.Date() + runif(1, min = -730, max = 0),
      bp_systolic = sample(30:160, 9, replace = FALSE),
      bp_diastolic = sample(10:100, 9, replace = FALSE),
      hba1c = sample(4:14, 9, replace = FALSE)
    ),
    time = c(rep("baseline", 3), rep("6mth", 3), rep("12mth", 3))
  ) %>%
    mutate(
      visit_date = case_when(
        time == "baseline" ~ enrolment_date + sample(-180:0, 3, replace = FALSE),
        time == "6mth" ~ enrolment_date + sample(100:180, 3, replace = FALSE),
        time == "12mth" ~ enrolment_date + sample(240:360, 3, replace = FALSE)
      ),
      days_difference = visit_date - enrolment_date
    ) %>%
    relocate(enrolment_date, .after = visit_date) %>%
    relocate(time, .after = days_difference)
}

set.seed(2024)
df_long <- rbind(
  generate_fake_patient_data("P01"),
  generate_fake_patient_data("P02"),
  generate_fake_patient_data("P03")
) %>%
  dplyr::arrange(patient_id, visit_date)

# Master dataframe
df_master <- df_long %>% distinct(patient_id, enrolment_date)
print(df_master)
#>   patient_id enrolment_date
#> 1        P01     2024-02-16
#> 2        P02     2023-06-08
#> 3        P03     2023-12-14

# Longitudinal data of readings
## 1. days_difference = visit_date - enrolment_date
## 2. enrolment_date, days_difference and time are not required as input but
##    are shown for illustration purpose
## 3. enrolment_date will be joined unto df_long from df_master internally
## 4. days_difference will be computed internally
print(df_long)
#>    patient_id bp_systolic bp_diastolic hba1c visit_date enrolment_date
#> 1         P01          74           38     4 2023-09-08     2024-02-16
#> 2         P01         138           71     6 2023-11-14     2024-02-16
#> 3         P01          66           25     8 2023-12-13     2024-02-16
#> 4         P01         134           43     9 2024-05-29     2024-02-16
#> 5         P01         151           35    11 2024-06-01     2024-02-16
#> 6         P01          61           23    13 2024-07-24     2024-02-16
#> 7         P01         127           53     7 2024-12-03     2024-02-16
#> 8         P01          58           59    14 2024-12-08     2024-02-16
#> 9         P01          40           95    12 2025-02-06     2024-02-16
#> 10        P02          96           12    14 2023-04-19     2023-06-08
#> 11        P02          54           34     8 2023-05-11     2023-06-08
#> 12        P02          63           74     5 2023-05-19     2023-06-08
#> 13        P02          86           29     4 2023-10-15     2023-06-08
#> 14        P02          78           18     6 2023-11-27     2023-06-08
#> 15        P02         118           15    12 2023-11-29     2023-06-08
#> 16        P02         119           86    10 2024-02-05     2023-06-08
#> 17        P02         155           89    11 2024-03-02     2023-06-08
#> 18        P02          83           88     7 2024-04-03     2023-06-08
#> 19        P03         107           43     7 2023-08-12     2023-12-14
#> 20        P03          45           28     4 2023-08-23     2023-12-14
#> 21        P03          49           27     9 2023-10-17     2023-12-14
#> 22        P03          65           21    11 2024-04-30     2023-12-14
#> 23        P03         110           75     6 2024-05-04     2023-12-14
#> 24        P03          75           90    10 2024-05-23     2023-12-14
#> 25        P03         101           25     8 2024-08-23     2023-12-14
#> 26        P03         160           71    14 2024-10-21     2023-12-14
#> 27        P03         102           54    13 2024-10-23     2023-12-14
#>    days_difference     time
#> 1        -161 days baseline
#> 2         -94 days baseline
#> 3         -65 days baseline
#> 4         103 days     6mth
#> 5         106 days     6mth
#> 6         159 days     6mth
#> 7         291 days    12mth
#> 8         296 days    12mth
#> 9         356 days    12mth
#> 10        -50 days baseline
#> 11        -28 days baseline
#> 12        -20 days baseline
#> 13        129 days     6mth
#> 14        172 days     6mth
#> 15        174 days     6mth
#> 16        242 days    12mth
#> 17        268 days    12mth
#> 18        300 days    12mth
#> 19       -124 days baseline
#> 20       -113 days baseline
#> 21        -58 days baseline
#> 22        138 days     6mth
#> 23        142 days     6mth
#> 24        161 days     6mth
#> 25        253 days    12mth
#> 26        312 days    12mth
#> 27        314 days    12mth

# Extract readings ----------------------------------------------------------
## 1. Baseline reading: latest reading in "-180 days <= days_difference <= 0 days"
## 2. 6-month reading: latest reading in "100 days <= days_difference <= 180 days"
## 3. 12-month reading: latest reading in "240 days <= days_difference <= 360 days"
df_master %>%
  # Blood pressure
  extract_reading_from_interval(
    list(
      ID = "patient_id",
      index_date = "enrolment_date"
    ),
    df_long,
    list(
      reading_date = "visit_date",
      reading = c("bp_systolic", "bp_diastolic")
    ),
    list(
      baseline = c(-180, 0),
      "6mth" = c(100, 180),
      "12mth" = c(240, 360)
    )
  ) %>%
  # HbA1c
  extract_reading_from_interval(
    list(
      ID = "patient_id",
      index_date = "enrolment_date"
    ),
    df_long,
    list(
      reading_date = "visit_date",
      reading = "hba1c"
    ),
    list(
      baseline = c(-180, 0),
      "6mth" = c(100, 180),
      "12mth" = c(240, 360)
    )
  ) %>%
  print()
#> <enrolment_date> is dropped from df_long,
#> instead it will be taken from df_master
#> <enrolment_date> is dropped from df_long,
#> instead it will be taken from df_master
#>   patient_id enrolment_date bp_systolic_date_baseline bp_systolic_baseline
#> 1        P01     2024-02-16                2023-12-13                   66
#> 2        P02     2023-06-08                2023-05-19                   63
#> 3        P03     2023-12-14                2023-10-17                   49
#>   bp_diastolic_baseline bp_systolic_date_6mth bp_systolic_6mth
#> 1                    25            2024-07-24               61
#> 2                    74            2023-11-29              118
#> 3                    27            2024-05-23               75
#>   bp_diastolic_6mth bp_systolic_date_12mth bp_systolic_12mth bp_diastolic_12mth
#> 1                23             2025-02-06                40                 95
#> 2                15             2024-04-03                83                 88
#> 3                90             2024-10-23               102                 54
#>   hba1c_date_baseline hba1c_baseline hba1c_date_6mth hba1c_6mth
#> 1          2023-12-13              8      2024-07-24         13
#> 2          2023-05-19              5      2023-11-29         12
#> 3          2023-10-17              9      2024-05-23         10
#>   hba1c_date_12mth hba1c_12mth
#> 1       2025-02-06          12
#> 2       2024-04-03           7
#> 3       2024-10-23          13