Save dataframe to an encrypted RDS file

ezyr_save_rds saves your dataframe to an RDS file and encrypts it with a sodium symmetric key.

Usage

ezyr_save_rds(df, filepath, key)

Arguments

df

(data.frame) A dataframe which you want to save to rds format

filepath

(character) File path to save your RDS file to

key

(cyphr_key) Sodium key obtained from invoking ask_for_key

Value

(data.frame) Original input dataframe

Examples

library(diffdf)

# Create random dataframe for illustration purpose
set.seed(123)
df <- data.frame(
  col1 = sample(letters, 10, replace = TRUE), # Random characters
  col2 = round(runif(10, min = 1, max = 100), 0), # Random integers between 1 and 100
  col3 = runif(10, min = 0, max = 1) # Random floats between 0 and 1
)

# Get a sodium key
key <- ask_for_key("my_password")

# Save RDS file in temp directory
filepath <- file.path(tempdir(), "file1.rds")
ezyr_save_rds(df, filepath, key)
print(list.files(tempdir(), pattern = "\\.rds$"))
#> [1] "file1.rds"

# Read the RDS file back
df_readback <- ezyr_read_rds(filepath, key)

# Compare the read back RDS file with original
diffdf(base = df, compare = df_readback)
#> No issues were found!

# Delete the saved RDS file in temp directory
unlink(filepath)
print(list.files(tempdir(), pattern = "\\.rds$"))
#> character(0)