Get presentation template for a generalised linear model

get_presentation_template_glm creates a data.frame template that lays out the variables in your model e.g. levels of a factor variable.

Usage

get_presentation_template_glm(model)

Arguments

model

(glm object) A stats::glm model object

Value

(data.frame) A template for you to left join unto to present your regression results

Details

get_presentation_template_glm has been incorporated into tabulate_glm_result, to create a table of your GLM results for presentation in an Rmarkdown document.

Examples

library(dplyr)
if (require("mlbench") & require("forcats")) {
  data(BostonHousing2, package = "mlbench")

  glm(medv ~ crim + zn + indus + town + rm + age + dis,
    data = BostonHousing2 %>%
      filter(town %in% c(
        "Newton", "Boston South Boston", "Boston Roxbury",
        "Somerville", "Boston Savin Hill", "Cambridge"
      )) %>%
      dplyr::mutate_at("town", forcats::fct_drop),
    family = gaussian(link = "identity")
  ) %>%
    get_presentation_template_glm()
}
#>          variable_name   var0                     var
#> 1                 crim header                    crim
#> 2                   zn header                      zn
#> 3                indus header                   indus
#> 4                 town header                    <NA>
#> 5       Boston Roxbury    Ref                    <NA>
#> 6    Boston Savin Hill   <NA>   townBoston Savin Hill
#> 7  Boston South Boston   <NA> townBoston South Boston
#> 8            Cambridge   <NA>           townCambridge
#> 9               Newton   <NA>              townNewton
#> 10          Somerville   <NA>          townSomerville
#> 11                  rm header                      rm
#> 12                 age header                     age
#> 13                 dis header                     dis