Get presentation template for a multinomial logistic regression model

get_presentation_template_mlogit creates a data.frame template that lays out the variables in your model.

Usage

get_presentation_template_mlogit(model, data_labels = NULL)

Arguments

model

(mlogit object) A mlogit::mlogit model object

data_labels

(list) A named list of labels

names: variable names

values: labels

If the labels were set with the labelled package, the list is obtained by invoking labelled::var_label()

Value

(data.frame) Table of your multinomial logistic regression results

Examples

library(mlogit)
#> Loading required package: dfidx
#> 
#> Attaching package: ‘dfidx’
#> The following object is masked from ‘package:stats’:
#> 
#>     filter
data(iris)

# Run a multinomial logistic regression on iris dataset ---------------------
mlogit(Species ~ 1 | Sepal.Length + Sepal.Width + Petal.Length + Petal.Width,
  data = mlogit.data(
    iris %>%
      mutate(Petal.Length = ifelse(
        Petal.Length < 2.5,
        "Petal length < 2.5",
        "Petal length >= 2.5"
      ) %>%
        forcats::fct_relevel("Petal length < 2.5", "Petal length >= 2.5")),
    choice = "Species",
    shape = "wide"
  ),
  reflevel = "setosa",
  na.action = na.omit
) %>%
  get_presentation_template_mlogit()
#>         variable_name   var0                             var
#> 1        Sepal.Length header                    Sepal.Length
#> 2         Sepal.Width header                     Sepal.Width
#> 3        Petal.Length header                            <NA>
#> 4  Petal length < 2.5    Ref                            <NA>
#> 5 Petal length >= 2.5   <NA> Petal.LengthPetal length >= 2.5
#> 6         Petal.Width header                     Petal.Width

# Run a multinomial logistic regression on labelled iris dataset ------------
labels <- list(
  Sepal.Length = "Sepal length",
  Sepal.Width = "Sepal width",
  Petal.Length = "Petal length",
  Petal.Width = "Petal width"
)
labelled::var_label(iris) <- labels
mlogit(Species ~ 1 | Sepal.Length + Sepal.Width + Petal.Length + Petal.Width,
  data = mlogit.data(
    iris %>%
      mutate(Petal.Length = ifelse(
        Petal.Length < 2.5,
        "Petal length < 2.5",
        "Petal length >= 2.5"
      ) %>%
        forcats::fct_relevel("Petal length < 2.5", "Petal length >= 2.5")),
    choice = "Species",
    shape = "wide"
  ),
  reflevel = "setosa",
  na.action = na.omit
) %>%
  get_presentation_template_mlogit(data_labels = labelled::var_label(iris))
#>         variable_name   var0                             var
#> 1        Sepal length header                    Sepal.Length
#> 2         Sepal width header                     Sepal.Width
#> 3        Petal length header                            <NA>
#> 4  Petal length < 2.5    Ref                            <NA>
#> 5 Petal length >= 2.5   <NA> Petal.LengthPetal length >= 2.5
#> 6         Petal width header                     Petal.Width