Safe left join

safe_left_join is a wrapper around dplyr::left_join that does an additional check for additional rows joined unto the left data frame.

Usage

safe_left_join(left_df, right_df, ...)

Arguments

left_df

(data.frame) Left data frame to left join to

right_df

(data.frame) Right data frame to join unto the left data frame

...

(ellipsis) Extra arguments to pass into dplyr::left_join

Value

(data.frame) left_df, the left data frame

Examples

# Left table ----------------------------------------------------------------
left_table <- data.frame(
  ID = sprintf(
    "%s%s",
    sample(1000:2000, 5, replace = TRUE),
    sample(letters, 5, replace = TRUE)
  ),
  column_b = runif(5, min = 0, max = 1),
  column_c = sample(letters, 5, replace = TRUE)
)
print(left_table)
#>      ID  column_b column_c
#> 1 1462p 0.1028646        q
#> 2 1040t 0.4348927        v
#> 3 1430f 0.9849570        r
#> 4 1089k 0.8930511        q
#> 5 1315h 0.8864691        b

# Right table ---------------------------------------------------------------
right_table <- data.frame(
  ID = left_table$ID,
  column_d = runif(5, min = 10, max = 50)
)
print(right_table)
#>      ID column_d
#> 1 1462p 22.81493
#> 2 1040t 17.50764
#> 3 1430f 41.29177
#> 4 1089k 13.74380
#> 5 1315h 28.67116

# Left table has same row count after join ----------------------------------
left_table %>%
  safe_left_join(right_table, by = "ID") %>%
  print()
#>      ID  column_b column_c column_d
#> 1 1462p 0.1028646        q 22.81493
#> 2 1040t 0.4348927        v 17.50764
#> 3 1430f 0.9849570        r 41.29177
#> 4 1089k 0.8930511        q 13.74380
#> 5 1315h 0.8864691        b 28.67116

# Right table with duplicated ID --------------------------------------------
right_table2 <- rbind(right_table, data.frame(
  ID = right_table[3:4, "ID"],
  column_d = runif(2, min = 10, max = 50)
))
print(right_table2)
#>      ID column_d
#> 1 1462p 22.81493
#> 2 1040t 17.50764
#> 3 1430f 41.29177
#> 4 1089k 13.74380
#> 5 1315h 28.67116
#> 6 1430f 30.46022
#> 7 1089k 33.99956

# Left table has more rows after join ---------------------------------------
try(
  left_table %>%
    safe_left_join(right_table2, by = "ID") %>%
    print()
)
#> Error : Row count after join not equals row count before join