Tabulate GLM result

tabulate_glm_result tabulates your GLM results for presentation in an Rmarkdown document.

Usage

tabulate_glm_result(model, conf_lvl = 0.95, exponentiate = FALSE, num_dp = 2)

Arguments

model

(glm object) A stats::glm model object

conf_lvl

(numeric) Confidence level. The default value of 0.95 for 95% confidence level.

By default, summary.glm function in R computes p-values using the Wald method. To be consistent, we compute confidence intervals using the Wald method via stats::confint.default. See discussion in this post

exponentiate

(logical) TRUE if beta-coefficients are to be exponentiated and expressed as odds ratios. If TRUE, confidence intervals will also be exponentiated correspondingly

num_dp

(numeric) Number of decimal places to present beta coefficients/odds ratios and confidence intervals

Value

(data.frame) Table of your GLM results

Examples

library(dplyr)
if (require("mlbench") & require("forcats")) {
  data(BostonHousing2, package = "mlbench")

  glm(medv ~ crim + town + rm + age + dis,
    data = BostonHousing2 %>%
      filter(town %in% c(
        "Newton", "Boston South Boston", "Boston Roxbury",
        "Somerville", "Boston Savin Hill", "Cambridge"
      )) %>%
      dplyr::mutate_at("town", forcats::fct_drop),
    family = gaussian(link = "identity")
  ) %>%
    tabulate_glm_result()
}
#>          variable_name   var0                     var        beta  std_error
#> 1                 crim header                    crim  0.03281571 0.04684588
#> 2                 town header                    <NA>          NA         NA
#> 3       Boston Roxbury    Ref                    <NA>          NA         NA
#> 4    Boston Savin Hill   <NA>   townBoston Savin Hill  0.15680129 1.51329455
#> 5  Boston South Boston   <NA> townBoston South Boston -0.34069764 1.67473462
#> 6            Cambridge   <NA>           townCambridge 13.31944678 1.56165139
#> 7               Newton   <NA>              townNewton  9.49983453 2.72208844
#> 8           Somerville   <NA>          townSomerville  7.47983762 1.81927838
#> 9                   rm header                      rm  9.23141134 0.68872082
#> 10                 age header                     age -0.09300762 0.04381661
#> 11                 dis header                     dis  0.69699530 1.58478814
#>       t_value      p_value   CI_lbound    CI_ubound          Beta (95% CI)
#> 1   0.7005037 4.851192e-01 -0.05900053  0.124631953   0.03 (-0.06 to 0.12)
#> 2          NA           NA          NA           NA                   <NA>
#> 3          NA           NA          NA           NA                   <NA>
#> 4   0.1036158 9.176664e-01 -2.80920153  3.122804112   0.16 (-2.81 to 3.12)
#> 5  -0.2034338 8.391788e-01 -3.62311718  2.941721896  -0.34 (-3.62 to 2.94)
#> 6   8.5290781 9.987464e-14 10.25866629 16.380227273 13.32 (10.26 to 16.38)
#> 7   3.4899066 7.006615e-04  4.16463924 14.835029830   9.50 (4.16 to 14.84)
#> 8   4.1114311 7.690232e-05  3.91411752 11.045557718   7.48 (3.91 to 11.05)
#> 9  13.4037060 1.022210e-24  7.88154334 10.581279333   9.23 (7.88 to 10.58)
#> 10 -2.1226567 3.606770e-02 -0.17888660 -0.007128644 -0.09 (-0.18 to -0.01)
#> 11  0.4398035 6.609583e-01 -2.40913238  3.803122970   0.70 (-2.41 to 3.80)
#>    p_value_
#> 1      0.49
#> 2      <NA>
#> 3      <NA>
#> 4      0.92
#> 5      0.84
#> 6    <0.001
#> 7    <0.001
#> 8    <0.001
#> 9    <0.001
#> 10     0.04
#> 11     0.66