Tabulate multinomial logistic regression result

tabulate_mlogit_result tabulates the results of your multinomial logistic regression that were created from running mlogit::mlogit.

Usage

tabulate_mlogit_result(
  model,
  outcome_levels_2onwards,
  data_labels = NULL,
  num_dp = 2
)

Arguments

model

(mlogit object) A mlogit::mlogit model object

outcome_levels_2onwards

(character) A character vector of outcome levels, from level 2 onwards

outcome level 1 refers to the reference level (omitted)

outcome levels 2 onwards refers to other outcome levels other than the reference level

the results for each of the outcome levels will be tabulated in order, from left to right

data_labels

(list) A named list of labels

names: variable names

values: labels

If the labels were set with the labelled package, the list is obtained by invoking labelled::var_label()

num_dp

(numeric) Number of decimal places to present confidence intervals. Defaults to 2 d.p.

Value

(data.frame) Table of your multinomial logistic regression results

Examples

library(mlogit)
data(iris)

# Run a multinomial logistic regression on iris dataset ---------------------
mlogit(Species ~ 1 | Sepal.Length + Sepal.Width + Petal.Length + Petal.Width,
  data = mlogit.data(
    iris %>%
      mutate(Petal.Length = ifelse(
        Petal.Length < 2.5,
        "Petal length < 2.5",
        "Petal length >= 2.5"
      ) %>%
        forcats::fct_relevel("Petal length < 2.5", "Petal length >= 2.5")),
    choice = "Species",
    shape = "wide"
  ),
  reflevel = "setosa",
  na.action = na.omit
) %>%
  tabulate_mlogit_result(c("versicolor", "virginica"))
#>         variable_name   var0                             var
#> 1        Sepal.Length header                    Sepal.Length
#> 2         Sepal.Width header                     Sepal.Width
#> 3        Petal.Length header                            <NA>
#> 4  Petal length < 2.5    Ref                            <NA>
#> 5 Petal length >= 2.5   <NA> Petal.LengthPetal length >= 2.5
#> 6         Petal.Width header                     Petal.Width
#>                                   Variable.x     Beta.x Std. Error.x
#> 1                    Sepal.Length:versicolor -0.5881255     5942.981
#> 2                     Sepal.Width:versicolor  2.4638702     7857.212
#> 3                                       <NA>         NA           NA
#> 4                                       <NA>         NA           NA
#> 5 Petal.LengthPetal length >= 2.5:versicolor 54.8338912    25059.441
#> 6                     Petal.Width:versicolor -8.9519115    13945.986
#>       z-value.x p_value.x CI_lower.x CI_upper.x odds_ratio.x    95% CI.x
#> 1 -9.896137e-05 0.9999210          0        Inf 5.553673e-01 0.00 to Inf
#> 2  3.135807e-04 0.9997498          0        Inf 1.175020e+01 0.00 to Inf
#> 3            NA        NA         NA         NA           NA        <NA>
#> 4            NA        NA         NA         NA           NA        <NA>
#> 5  2.188153e-03 0.9982541          0        Inf 6.517130e+23 0.00 to Inf
#> 6 -6.418988e-04 0.9994878          0        Inf 1.294894e-04 0.00 to Inf
#>   p_value_.x                                Variable.y     Beta.y Std. Error.y
#> 1       1.00                    Sepal.Length:virginica  0.7070037     5942.981
#> 2       1.00                     Sepal.Width:virginica -2.3594039     7857.212
#> 3       <NA>                                      <NA>         NA           NA
#> 4       <NA>                                      <NA>         NA           NA
#> 5       1.00 Petal.LengthPetal length >= 2.5:virginica 28.3797729    21485.755
#> 6       1.00                     Petal.Width:virginica  6.9707490    13945.986
#>       z-value.y p_value.y CI_lower.y CI_upper.y odds_ratio.y    95% CI.y
#> 1  0.0001189645 0.9999051          0        Inf 2.027906e+00 0.00 to Inf
#> 2 -0.0003002851 0.9997604          0        Inf 9.447652e-02 0.00 to Inf
#> 3            NA        NA         NA         NA           NA        <NA>
#> 4            NA        NA         NA         NA           NA        <NA>
#> 5  0.0013208646 0.9989461          0        Inf 2.114359e+12 0.00 to Inf
#> 6  0.0004998391 0.9996012          0        Inf 1.065020e+03 0.00 to Inf
#>   p_value_.y
#> 1       1.00
#> 2       1.00
#> 3       <NA>
#> 4       <NA>
#> 5       1.00
#> 6       1.00

# Run a multinomial logistic regression on labelled iris dataset ------------
labels <- list(
  Sepal.Length = "Sepal length",
  Sepal.Width = "Sepal width",
  Petal.Length = "Petal length",
  Petal.Width = "Petal width"
)
labelled::var_label(iris) <- labels
mlogit(Species ~ 1 | Sepal.Length + Sepal.Width + Petal.Length + Petal.Width,
  data = mlogit.data(
    iris %>%
      mutate(Petal.Length = ifelse(
        Petal.Length < 2.5,
        "Petal length < 2.5",
        "Petal length >= 2.5"
      ) %>%
        forcats::fct_relevel("Petal length < 2.5", "Petal length >= 2.5")),
    choice = "Species",
    shape = "wide"
  ),
  reflevel = "setosa",
  na.action = na.omit
) %>%
  tabulate_mlogit_result(c("versicolor", "virginica"), data_labels = labelled::var_label(iris))
#>         variable_name   var0                             var
#> 1        Sepal length header                    Sepal.Length
#> 2         Sepal width header                     Sepal.Width
#> 3        Petal length header                            <NA>
#> 4  Petal length < 2.5    Ref                            <NA>
#> 5 Petal length >= 2.5   <NA> Petal.LengthPetal length >= 2.5
#> 6         Petal width header                     Petal.Width
#>                                   Variable.x     Beta.x Std. Error.x
#> 1                    Sepal.Length:versicolor -0.5881255     5942.981
#> 2                     Sepal.Width:versicolor  2.4638702     7857.212
#> 3                                       <NA>         NA           NA
#> 4                                       <NA>         NA           NA
#> 5 Petal.LengthPetal length >= 2.5:versicolor 54.8338912    25059.441
#> 6                     Petal.Width:versicolor -8.9519115    13945.986
#>       z-value.x p_value.x CI_lower.x CI_upper.x odds_ratio.x    95% CI.x
#> 1 -9.896137e-05 0.9999210          0        Inf 5.553673e-01 0.00 to Inf
#> 2  3.135807e-04 0.9997498          0        Inf 1.175020e+01 0.00 to Inf
#> 3            NA        NA         NA         NA           NA        <NA>
#> 4            NA        NA         NA         NA           NA        <NA>
#> 5  2.188153e-03 0.9982541          0        Inf 6.517130e+23 0.00 to Inf
#> 6 -6.418988e-04 0.9994878          0        Inf 1.294894e-04 0.00 to Inf
#>   p_value_.x                                Variable.y     Beta.y Std. Error.y
#> 1       1.00                    Sepal.Length:virginica  0.7070037     5942.981
#> 2       1.00                     Sepal.Width:virginica -2.3594039     7857.212
#> 3       <NA>                                      <NA>         NA           NA
#> 4       <NA>                                      <NA>         NA           NA
#> 5       1.00 Petal.LengthPetal length >= 2.5:virginica 28.3797729    21485.755
#> 6       1.00                     Petal.Width:virginica  6.9707490    13945.986
#>       z-value.y p_value.y CI_lower.y CI_upper.y odds_ratio.y    95% CI.y
#> 1  0.0001189645 0.9999051          0        Inf 2.027906e+00 0.00 to Inf
#> 2 -0.0003002851 0.9997604          0        Inf 9.447652e-02 0.00 to Inf
#> 3            NA        NA         NA         NA           NA        <NA>
#> 4            NA        NA         NA         NA           NA        <NA>
#> 5  0.0013208646 0.9989461          0        Inf 2.114359e+12 0.00 to Inf
#> 6  0.0004998391 0.9996012          0        Inf 1.065020e+03 0.00 to Inf
#>   p_value_.y
#> 1       1.00
#> 2       1.00
#> 3       <NA>
#> 4       <NA>
#> 5       1.00
#> 6       1.00