Label prefixed or suffixed dataframe columns

label_ix_columns adds labels to column(s) of your data.frame which have the particular prefix/suffix you specified, by copying the labels from corresponding columns without the prefix/suffix. The labels are added with labelled::var_label().

Usage

label_ix_columns(df, pattern, suffix = TRUE)

Arguments

df

(data.frame) Your data in the form of a data frame

pattern

(character) regex pattern to identify your prefix/suffix

suffix

(logical) TRUE if the pattern is a suffix. FALSE if the pattern is a prefix

Value

(data.frame) Your data with new labels added

Examples

library(labelled)

# Example of suffix ---------------------------------------------------------
df <- data.frame(
  sex = sample(1:2, 5, replace = TRUE),
  race = sample(1:4, 5, replace = TRUE)
)
df$sex.factor <- c("Female", "Male")[df$sex]
df$race.factor <- c("Chinese", "Malay", "Indian", "Others")[df$race]

# Set labels
labels <- list(
  sex = "Sex",
  race = "Race"
)
var_label(df) <- labels

# Get labels after label_ix_columns()
var_label(label_ix_columns(df, ".factor", suffix = TRUE))
#> $sex
#> [1] "Sex"
#> 
#> $race
#> [1] "Race"
#> 
#> $sex.factor
#> [1] "Sex"
#> 
#> $race.factor
#> [1] "Race"
#> 

# Example of prefix ---------------------------------------------------------
df <- data.frame(
  sex = sample(1:2, 5, replace = TRUE),
  race = sample(1:4, 5, replace = TRUE)
)
df$fct_sex <- c("Female", "Male")[df$sex]
df$fct_race <- c("Chinese", "Malay", "Indian", "Others")[df$race]

# Set labels
labels <- list(
  sex = "Sex",
  race = "Race"
)
var_label(df) <- labels

# Get labels after label_ix_columns()
var_label(label_ix_columns(df, "fct_", suffix = FALSE))
#> $sex
#> [1] "Sex"
#> 
#> $race
#> [1] "Race"
#> 
#> $fct_sex
#> [1] "Sex"
#> 
#> $fct_race
#> [1] "Race"
#>