Safe left join

safe_left_join is a wrapper around dplyr::left_join that does an additional check for additional rows joined unto the left data frame.

safe_left_join(left_df, right_df, ...)

Arguments

left_df

(data.frame) Left data frame to left join to

right_df

(data.frame) Right data frame to join unto the left data frame

...

(ellipsis) Extra arguments to pass into dplyr::left_join

Value

(data.frame) left_df, the left data frame

Examples

# Left table ----------------------------------------------------------------
left_table <- data.frame(
  ID = sprintf(
    "%s%s",
    sample(1000:2000, 5, replace = TRUE),
    sample(letters, 5, replace = TRUE)
  ),
  column_b = runif(5, min = 0, max = 1),
  column_c = sample(letters, 5, replace = TRUE)
)
print(left_table)
#>      ID  column_b column_c
#> 1 1772f 0.7064338        u
#> 2 1468d 0.9485766        f
#> 3 1299p 0.1803388        i
#> 4 1956b 0.2168999        v
#> 5 1545c 0.6801629        j

# Right table ---------------------------------------------------------------
right_table <- data.frame(
  ID = left_table$ID,
  column_d = runif(5, min = 10, max = 50)
)
print(right_table)
#>      ID column_d
#> 1 1772f 49.62849
#> 2 1468d 48.82084
#> 3 1299p 25.56731
#> 4 1956b 28.44746
#> 5 1545c 22.60967

# Left table has same row count after join ----------------------------------
left_table %>%
  safe_left_join(right_table, by = "ID") %>%
  print()
#>      ID  column_b column_c column_d
#> 1 1772f 0.7064338        u 49.62849
#> 2 1468d 0.9485766        f 48.82084
#> 3 1299p 0.1803388        i 25.56731
#> 4 1956b 0.2168999        v 28.44746
#> 5 1545c 0.6801629        j 22.60967

# Right table with duplicated ID --------------------------------------------
right_table2 <- rbind(right_table, data.frame(
  ID = right_table[3:4, "ID"],
  column_d = runif(2, min = 10, max = 50)
))
print(right_table2)
#>      ID column_d
#> 1 1772f 49.62849
#> 2 1468d 48.82084
#> 3 1299p 25.56731
#> 4 1956b 28.44746
#> 5 1545c 22.60967
#> 6 1299p 16.98704
#> 7 1956b 31.26294

# Left table has more rows after join ---------------------------------------
try(
  left_table %>%
    safe_left_join(right_table2, by = "ID") %>%
    print()
)
#> Error : Row count after join not equals row count before join